refactor(niivue): share channel-file assembly and typed-array allocation
The origin-centred affine plus NIfTI File construction existed three
times (Allen atlas, OME-TIFF, OME-Zarr channel loaders); all three now
delegate to channelVolumeFile so the convention cannot drift. The
allocLike helper duplicated across omeZarrLoader and omeZarrChunkedSource
becomes allocTypedArrayLike in omeZarr.ts. The chunk source's clamp/pad
repack and the loader's whole-block transpose stay separate on purpose:
they are unlike operations and forcing them into one abstraction would
obscure the padding contract.
